Let the dead speak

+++ UnityProject/Assets/Scripts/Messages/Client/PostToChatMessage.cs
@@ -58,7 +58,7 @@ public class PostToChatMessage : ClientMessage<PostToChatMessage>
 	public override void Deserialize(NetworkReader reader)
 	{
 		base.Deserialize(reader);
-		Channels = (ChatChannel)reader.ReadInt16();
+		Channels = (ChatChannel)reader.ReadUInt32();
 		ChatMessageText = reader.ReadString();
 
 	}
@@ -66,7 +66,7 @@ public class PostToChatMessage : ClientMessage<PostToChatMessage>
 	public override void Serialize(NetworkWriter writer)
 	{
 		base.Serialize(writer);
-		writer.Write((Int16)Channels);
+		writer.Write((Int32)Channels);
 		writer.Write(ChatMessageText);
 	}
 }
